ocean-plotter-ows/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.9.5
aiosignal==1.3.1
asciitree==0.3.3
attrs==24.2.0
basemap==1.4.1
basemap-data==1.3.2
basemap-data-hires==1.3.2
beautifulsoup4==4.12.3
boto3==1.35.78
botocore==1.35.78
cachier==2.3.0
certifi==2024.8.30
cftime==1.6.4.post1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.1
copernicusmarine==1.3.5
cycler==0.12.1
dask==2022.12.1
dill==0.3.9
fasteners==0.19
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
geopandas==1.0.1
idna==3.10
jmespath==1.0.1
kiwisolver==1.4.7
locket==1.0.0
lxml==4.9.4
matplotlib==3.8.4
motuclient==1.8.8
multidict==6.1.0
multiprocess==0.70.17
nest-asyncio==1.6.0
netCDF4==1.7.2
numcodecs==0.14.1
numpy==1.26.4
OWSLib==0.32.0
packaging==23.2
pandas==2.2.3
partd==1.4.2
pathos==0.3.3
pillow==11.0.0
portalocker==3.0.0
pox==0.3.5
ppft==1.7.6.9
propcache==0.2.1
pydap==3.5.2
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.6.1
pyshp==2.3.1
pystac==1.11.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
scipy==1.14.1
semver==3.0.2
setuptools==68.2.2
shapely==2.0.7
six==1.17.0
soupsieve==2.6
toolz==1.0.0
tqdm==4.67.1
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
WebOb==1.8.9
xarray==2023.12.0
yarl==1.18.3
zarr==2.18.3
